+++ src/Loop/Exception/RunningError.php
@@ -11,8 +11,11 @@ namespace Icicle\Loop\Exception;
 
 class RunningError extends \Error implements Error
 {
-    public function __construct()
+    /**
+     * @param string|null $message
+     */
+    public function __construct(string $message = null)
     {
-        parent::__construct('The loop was already running.');
+        parent::__construct($message ?: 'The loop was already running.');
     }
 }