+++ src/Loop/Manager/Ev/EvIoManager.php
@@ -57,9 +57,15 @@ class EvIoManager implements IoManager
             $io = $event->data;
             $id = (int) $io->getResource();
 
-            $event->stop();
-            if (isset($this->timers[$id])) {
-                $this->timers[$id]->stop();
+            if ($io->isPersistent()) {
+                if (isset($this->timers[$id])) {
+                    $this->timers[$id]->again();
+                }
+            } else {
+                $event->stop();
+                if (isset($this->timers[$id])) {
+                    $this->timers[$id]->stop();
+                }
             }
 
             $io->call(false);
@@ -108,7 +114,7 @@ class EvIoManager implements IoManager
     /**
      * {@inheritdoc}
      */
-    public function create($resource, callable $callback): Io
+    public function create($resource, callable $callback, bool $persistent = false): Io
     {
         $id = (int) $resource;
         
@@ -116,7 +122,7 @@ class EvIoManager implements IoManager
             throw new ResourceBusyError();
         }
 
-        $socket = new Io($this, $resource, $callback);
+        $socket = new Io($this, $resource, $callback, $persistent);
 
         $event = $this->loop->io($resource, $this->type, $this->socketCallback, $socket);
         $event->stop();
@@ -145,10 +151,10 @@ class EvIoManager implements IoManager
             }
 
             if (isset($this->timers[$id])) {
-                $this->timers[$id]->set($timeout, 0);
+                $this->timers[$id]->set($timeout, $timeout);
                 $this->timers[$id]->start();
             } else {
-                $this->timers[$id] = $this->loop->timer($timeout, 0, $this->timerCallback, $io);
+                $this->timers[$id] = $this->loop->timer($timeout, $timeout, $this->timerCallback, $io);
             }
         }
     }