+++ src/Loop/Manager/Select/SelectIoManager.php
@@ -65,7 +65,7 @@ class SelectIoManager implements IoManager
     /**
      * {@inheritdoc}
      */
-    public function create($resource, callable $callback): Io
+    public function create($resource, callable $callback, bool $persistent = false): Io
     {
         $id = (int) $resource;
         
@@ -73,7 +73,7 @@ class SelectIoManager implements IoManager
             throw new ResourceBusyError();
         }
         
-        return $this->sockets[$id] = new Io($this, $resource, $callback);
+        return $this->sockets[$id] = new Io($this, $resource, $callback, $persistent);
     }
     
     /**
@@ -171,12 +171,16 @@ class SelectIoManager implements IoManager
     {
         foreach ($active as $id => $resource) {
             if (isset($this->sockets[$id], $this->pending[$id])) { // Event may have been removed from a previous call.
-                unset($this->pending[$id]);
-                
-                if (isset($this->timers[$id])) {
-                    $this->timers[$id]->stop();
+
+                if (!$this->sockets[$id]->isPersistent()) {
+                    unset($this->pending[$id]);
+                    if (isset($this->timers[$id])) {
+                        $this->timers[$id]->stop();
+                    }
+                } elseif (isset($this->timers[$id])) {
+                    $this->timers[$id]->again();
                 }
-                
+
                 $this->sockets[$id]->call(false);
             }
         }