+++ src/Loop/functions.php
@@ -152,23 +152,25 @@ if (!function_exists(__NAMESPACE__ . '\loop')) {
     /**
      * @param resource $socket Stream socket resource.
      * @param callable $callback Callback to be invoked when data is available on the socket.
+     * @param bool $persistent
      *
      * @return \Icicle\Loop\Watcher\Io
      */
-    function poll($socket, callable $callback): Io
+    function poll($socket, callable $callback, bool $persistent = false): Io
     {
-        return loop()->poll($socket, $callback);
+        return loop()->poll($socket, $callback, $persistent);
     }
 
     /**
      * @param resource $socket Stream socket resource.
      * @param callable $callback Callback to be invoked when the socket is available to write.
+     * @param bool $persistent
      *
      * @return \Icicle\Loop\Watcher\Io
      */
-    function await($socket, callable $callback): Io
+    function await($socket, callable $callback, bool $persistent = false): Io
     {
-        return loop()->await($socket, $callback);
+        return loop()->await($socket, $callback, $persistent);
     }
 
     /**